# output/final_deliverables/openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty100 Analytics API","description":"REST API for Nifty100 financial analytics and screening.","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Return API health status and database row counts.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies":{"get":{"tags":["Companies"],"summary":"Get Companies","description":"Return the 92-company Sprint 6 universe with optional filters.","operationId":"get_companies_api_v1_companies_get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"market_cap_category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market Cap Category"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}":{"get":{"tags":["Companies"],"summary":"Get Company","description":"Return the company profile, latest KPIs, and sector information.","operationId":"get_company_api_v1_companies__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/pl":{"get":{"tags":["Companies"],"summary":"Get Company Pl","description":"Return P&L history with optional year filters.","operationId":"get_company_pl_api_v1_companies__ticker__pl_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/bs":{"get":{"tags":["Companies"],"summary":"Get Company Bs","description":"Return balance-sheet history with optional year filters.","operationId":"get_company_bs_api_v1_companies__ticker__bs_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/cashflow":{"get":{"tags":["Companies"],"summary":"Get Company Cashflow","description":"Return cash-flow history with optional year filters.","operationId":"get_company_cashflow_api_v1_companies__ticker__cashflow_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/ratios":{"get":{"tags":["Companies"],"summary":"Get Company Ratios","description":"Return financial-ratio history with an optional year filter.","operationId":"get_company_ratios_api_v1_companies__ticker__ratios_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/tearsheet":{"get":{"tags":["Companies"],"summary":"Get Company Tearsheet","description":"Return the pre-generated company tearsheet PDF.","operationId":"get_company_tearsheet_api_v1_companies__ticker__tearsheet_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener":{"get":{"tags":["Screener"],"summary":"Run Screener","description":"Return ranked companies matching the supplied screening filters.","operationId":"run_screener_api_v1_screener_get","parameters":[{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Min Roe"}},{"name":"max_de","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Max De"}},{"name":"min_fcf","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Min Fcf"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"min_rev_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Min Rev Cagr 5Yr"}},{"name":"min_pat_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Min Pat Cagr 5Yr"}},{"name":"max_pe","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Max Pe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors":{"get":{"tags":["Sectors"],"summary":"Get Sectors","description":"Return sector-level company counts and median latest-year KPIs.","operationId":"get_sectors_api_v1_sectors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sectors/{sector}/companies":{"get":{"tags":["Sectors"],"summary":"Get Sector Companies","description":"Return companies and latest KPIs for a sector.","operationId":"get_sector_companies_api_v1_sectors__sector__companies_get","parameters":[{"name":"sector","in":"path","required":true,"schema":{"type":"string","title":"Sector"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/peers/{group_name}":{"get":{"tags":["Peers"],"summary":"Get Peer Group","description":"Return companies in a peer group with latest percentile metrics.","operationId":"get_peer_group_api_v1_peers__group_name__get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/peers/compare":{"get":{"tags":["Peers"],"summary":"Compare Company With Peers","description":"Return company, peer-average, and benchmark radar data.","operationId":"compare_company_with_peers_api_v1_companies__ticker__peers_compare_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market-cap/{ticker}":{"get":{"tags":["Valuation"],"summary":"Get Market Cap History","description":"Return historical market-cap and valuation multiples for a company.","operationId":"get_market_cap_history_api_v1_market_cap__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/stats":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio Stats","description":"Return the portfolio KPI percentile statistics.","operationId":"get_portfolio_stats_api_v1_portfolio_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/{ticker}/documents":{"get":{"tags":["Documents"],"summary":"Get Company Documents","description":"Return annual-report documents and URL validity information.","operationId":"get_company_documents_api_v1_companies__ticker__documents_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Return basic API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}